improved tests
